server/requirements.txt
fastapi[standard]
uvicorn[standard]
sqlmodel
pillow
boto3
python-magic
aiofiles